+++ js/jquery.terminal.js
@@ -4,7 +4,7 @@
  *  __ / // // // // // _  // _// // / / // _  // _//     // //  \/ // _ \/ /
  * /  / // // // // // ___// / / // / / // ___// / / / / // // /\  // // / /__
  * \___//____ \\___//____//_/ _\_  / /_//____//_/ /_/ /_//_//_/ /_/ \__\_\___/
- *           \/              /____/                              version 2.17.5
+ *           \/              /____/                              version DEV
  *
  * This file is part of jQuery Terminal. https://terminal.jcubic.pl
  *
@@ -41,7 +41,7 @@
  *
  * broken image by Sophia Bai from the Noun Project (CC-BY)
  *
- * Date: Sat, 04 Jul 2020 13:40:56 +0000
+ * Date: Mon, 06 Jul 2020 13:08:04 +0000
  */
 /* global define, Map */
 /* eslint-disable */
@@ -3564,7 +3564,8 @@
                 // event in input with e.fake == true
                 return;
             }
-            if (!e.fake) {
+            // Meta is special keydown triggered by Emoji picker on Windows 10
+            if (!e.fake && e.key.toLowerCase() !== 'meta') {
                 no_keydown = false;
             }
             no_keypress = true;
@@ -4435,8 +4436,8 @@
     }
     // -------------------------------------------------------------------------
     $.terminal = {
-        version: '2.17.5',
-        date: 'Sat, 04 Jul 2020 13:40:56 +0000',
+        version: 'DEV',
+        date: 'Mon, 06 Jul 2020 13:08:04 +0000',
         // colors from https://www.w3.org/wiki/CSS/Properties/color/keywords
         color_names: [
             'transparent', 'currentcolor', 'black', 'silver', 'gray', 'white',